Interview preperation/requirements.txt
streamlit
openai
python-dotenv
PyPDF2
google-generativeai